Remove response cache hit bit on Q/R sig flags.
It doesn't really fit there. Whether the application can produce it has nothing to do with the rest of the flags. It's not readily available from dnstap (so where will it come from?), and we might want to know a lot more about the query servicing (e.g. backend type, whether any TTL special processing happened). I don't know enough to speculate on contents, and whether it should be in the query signature or the Q/R. So this should be punted to implementation-specific for now until there is experience and consensus.