Attempt at tackling the POSIX time question.
Add reference to POSIX time, and note that it does not include leap seconds. Add implementation section with slightly longer discussion of the issue, and just note that implementations may want to compensate.

Add an explicit CBOR subheading under implementation guidelines.

Also note that the earliest-time is another item that is conditionally mandatory.